Create styleArticle
Here, I made a new style for the article, I set the width to 500px, and set the padding of the articles to 20px. I learned from this that article formatting is one of those elements you put up at the top when designing blogs because it will make it so all the articles you post follow your set parameters.  This is a lot like defining your variables in iOS code at the beginning, so that your functions run smoothly and similarly later.